refactor: move RootNode, Root and RootLink into separate root module
