+++ oscars/src/collectors/mark_sweep_branded/tests/uaf.rs
@@ -22,7 +22,7 @@ fn test_uaf() {
     with_gc(|cx| {
         let dropped = Cell::new(false);
         cx.mutate(|mcx| {
-            let _gc = mcx.alloc(DetectDrop(&dropped)).unwrap();
+            let _gc = mcx.try_alloc(DetectDrop(&dropped)).unwrap();
         });
         cx.collect(); // Garbage collects 'gc' because it isn't rooted!
         assert!(dropped.get(), "It wasn't collected!");
