+++ oscars/src/collectors/mark_sweep_branded/tests/ui/root_cross_context.rs
@@ -13,7 +13,7 @@ fn main() {
     with_gc(|ctx1| {
         with_gc(|ctx2| {
             // root carries 'id of ctx1
-            let root = ctx1.mutate(|cx| cx.root(cx.alloc(123i32).unwrap()));
+            let root = ctx1.mutate(|cx| cx.root(cx.try_alloc(123i32).unwrap()).unwrap());
 
             ctx2.mutate(|cx| {
                 // ERROR: `'id` of `root` (ctx1) != `'id` of `cx` (ctx2)
